requirements.txt
appscript==1.3.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.1
django-cors-headers==4.4.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
idna==3.10
lxml==5.3.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
xlwings==0.33.3
